renamed to AutoComplete for better review diff
