+++ search-ui/src/components/search/AutoComplete.vue
